rom-planner-app/server/backend/requirements.txt
# backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
psycopg2-binary==2.9.9 # PostgreSQL adapter
openpyxl==3.1.2       # For Excel file generation
python-dotenv==1.0.1  # For reading .env files locally (though Podman Compose handles it)